requirements.txt
streamlit>=1.32.0
firebase-admin>=6.4.0
pandas>=2.0.0
fpdf2>=2.7.9
protobuf>=4.25.0
